Moved to policy/tf_policy and policy/tf_policy_template
